backend/requirements.txt
alembic==1.14.1
aniso8601==10.0.1
anyio==4.5.2
APScheduler==3.11.0
attrs==25.3.0
backports.zoneinfo==0.2.1
bcrypt==4.3.0
blinker==1.8.2
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
cloudinary==1.44.1
cryptography==46.0.2
distlib==0.3.9
exceptiongroup==1.3.0
Faker==35.2.2
filelock==3.16.1
firebase-admin==6.9.0
flasgger==0.9.7.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.16.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.10
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.1
mistune==3.1.3
msgpack==1.1.1
packaging==25.0
Pillow==9.5.0
pipenv==2024.4.1
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydub==0.25.1
PyJWT==2.9.0
pyotp==2.9.0
pyparsing==3.1.4
PyPDF2==3.0.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
reportlab==4.4.3
requests==2.31.0
resend==0.6.0
rpds-py==0.20.1
rsa==4.9.1
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
SQLAlchemy-serializer==1.4.12
tomli==2.2.1
typing_extensions==4.13.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.30.0
Werkzeug==3.0.6
zipp==3.20.2
